+++ valhalla/loki/tiles.h
@@ -9,8 +9,11 @@
 #include "baldr/edgeinfo.h"
 #include "baldr/nodeinfo.h"
 #include "baldr/traffictile.h"
+#include "proto/incidents.pb.h"
+#include "proto_conversions.h"
 
 #include <vtzero/builder.hpp>
+#include <vtzero/types.hpp>
 
 #include <cassert>
 #include <random>
@@ -50,6 +53,47 @@ protected:
   vtzero::index_value key_modes_;
   vtzero::index_value key_except_destination_;
   vtzero::index_value key_value_;
+};
+
+struct IncidentLayersBuilder;
+struct IncidentsAttributeTile {
+  const char* key_name;
+  std::string_view attribute_flag;
+  vtzero::index_value IncidentLayersBuilder::*key_member;
+
+  using value_func_t = vtzero::encoded_property_value (*)(const IncidentsTile::Metadata&);
+  value_func_t value_func;
+};
+
+class IncidentLayersBuilder : public vtzero::layer_builder {
+public:
+  explicit IncidentLayersBuilder(vtzero::tile_builder& tile,
+                                 const char* name,
+                                 const baldr::AttributesController& controller);
+
+  void add_point_feature(const IncidentsTile::Metadata&, const vtzero::point&);
+  void add_line_feature(const IncidentsTile::Metadata&, const std::vector<vtzero::point>& geometry);
+  vtzero::index_value key_type_;
+  vtzero::index_value key_impact_;
+  vtzero::index_value key_description_;
+  vtzero::index_value key_sub_type_;
+  vtzero::index_value key_sub_type_description_;
+  vtzero::index_value key_start_time_;
+  vtzero::index_value key_end_time_;
+  vtzero::index_value key_road_closed_;
+  vtzero::index_value key_congestion_value_;
+  vtzero::index_value key_lanes_blocked_;
+  vtzero::index_value key_creation_time_;
+  vtzero::index_value key_long_description_;
+  vtzero::index_value key_clear_lanes_;
+  vtzero::index_value key_num_lanes_blocked_;
+  vtzero::index_value key_length_;
+  vtzero::index_value key_id_;
+  vtzero::index_value key_iso_3166_1_alpha2_;
+  vtzero::index_value key_iso_3166_1_alpha3_;
+
+protected:
+  const baldr::AttributesController controller_;
 };
 
 /**
@@ -1397,6 +1441,145 @@ static constexpr NodeAttributeTile kNodeAttributes[] = {
     },
 };
 
+static constexpr IncidentsAttributeTile kIncidentAttributes[] = {
+    {
+        "description",
+        baldr::kIncidentDescription,
+        &IncidentLayersBuilder::key_description_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.description());
+        },
+    },
+    {
+        "type",
+        baldr::kIncidentType,
+        &IncidentLayersBuilder::key_type_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(
+              static_cast<std::string>(incidentTypeToString(meta.type())));
+        },
+    },
+    {
+        "impact",
+        baldr::kIncidentImpact,
+        &IncidentLayersBuilder::key_impact_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(
+              static_cast<std::string>(incidentImpactToString(meta.impact())));
+        },
+    },
+    {
+        "sub_type",
+        baldr::kIncidentSubType,
+        &IncidentLayersBuilder::key_sub_type_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.sub_type());
+        },
+    },
+    {
+        "sub_type_description",
+        baldr::kIncidentSubTypeDescription,
+        &IncidentLayersBuilder::key_sub_type_description_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.sub_type_description());
+        },
+    },
+    {
+        "start_time",
+        baldr::kIncidentStartTime,
+        &IncidentLayersBuilder::key_start_time_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.start_time());
+        },
+    },
+    {
+        "end_time",
+        baldr::kIncidentEndTime,
+        &IncidentLayersBuilder::key_end_time_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.end_time());
+        },
+    },
+    {
+        "road_closed",
+        baldr::kIncidentRoadClosed,
+        &IncidentLayersBuilder::key_road_closed_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.road_closed());
+        },
+    },
+    {
+        "congestion_value",
+        baldr::kIncidentCongestionValue,
+        &IncidentLayersBuilder::key_congestion_value_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.congestion().value());
+        },
+    },
+    {
+        "creation_time",
+        baldr::kIncidentCreationTime,
+        &IncidentLayersBuilder::key_creation_time_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.creation_time());
+        },
+    },
+    {
+        "long_description",
+        baldr::kIncidentLongDescription,
+        &IncidentLayersBuilder::key_long_description_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.long_description());
+        },
+    },
+    {
+        "clear_lanes",
+        baldr::kIncidentClearLanes,
+        &IncidentLayersBuilder::key_clear_lanes_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.clear_lanes());
+        },
+    },
+    {
+        "num_lanes_blocked",
+        baldr::kIncidentNumLanesBlocked,
+        &IncidentLayersBuilder::key_num_lanes_blocked_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.num_lanes_blocked());
+        },
+    },
+    {
+        "length",
+        baldr::kIncidentLength,
+        &IncidentLayersBuilder::key_length_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.length());
+        },
+    },
+    {
+        "id",
+        baldr::kIncidentId,
+        &IncidentLayersBuilder::key_id_,
+        [](const IncidentsTile::Metadata& meta) { return vtzero::encoded_property_value(meta.id()); },
+    },
+    {
+        "iso_3166_1_alpha2",
+        baldr::kIncidentIso31661Alpha2,
+        &IncidentLayersBuilder::key_iso_3166_1_alpha2_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.iso_3166_1_alpha2());
+        },
+    },
+    {
+        "iso_3166_1_alpha3",
+        baldr::kIncidentIso31661Alpha3,
+        &IncidentLayersBuilder::key_iso_3166_1_alpha3_,
+        [](const IncidentsTile::Metadata& meta) {
+          return vtzero::encoded_property_value(meta.iso_3166_1_alpha3());
+        },
+    },
+};
+
 // map from MVT prop name to controller attribute flag for edge properties
 static const std::unordered_map<std::string_view, std::string_view> kEdgePropToAttributeFlag = {
     // Forward edge attributes
@@ -1501,6 +1684,28 @@ static const std::unordered_map<std::string_view, std::string_view> kNodePropToA
     {"iso_3166_1", baldr::kAdminCountryCode},
     {"iso_3166_2", baldr::kAdminStateCode},
 };
+
+static const std::unordered_map<std::string_view, std::string_view> kIncidentPropToAttributeFlag = {
+    // incident
+    {"description", baldr::kIncidentDescription},
+    {"type", baldr::kIncidentType},
+    {"impact", baldr::kIncidentImpact},
+    {"sub_type", baldr::kIncidentSubType},
+    {"sub_type_description", baldr::kIncidentSubTypeDescription},
+    {"start_time", baldr::kIncidentStartTime},
+    {"end_time", baldr::kIncidentEndTime},
+    {"road_closed", baldr::kIncidentRoadClosed},
+    {"congestion_value", baldr::kIncidentCongestionValue},
+    {"creation_time", baldr::kIncidentCreationTime},
+    {"long_description", baldr::kIncidentLongDescription},
+    {"clear_lanes", baldr::kIncidentClearLanes},
+    {"num_lanes_blocked", baldr::kIncidentNumLanesBlocked},
+    {"length", baldr::kIncidentLength},
+    {"id", baldr::kIncidentId},
+    {"iso_3166_1_alpha2", baldr::kIncidentIso31661Alpha2},
+    {"iso_3166_1_alpha3", baldr::kIncidentIso31661Alpha3},
+};
+
 } // namespace detail
 
 } // namespace valhalla::loki